requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
opencv-python==4.5.3.20210927
pytesseract==0.3.8
tensorflow==2.6.0
firebase-admin==5.2.0
numpy==1.21.2
Pillow==8.4.0
Flask
flask-cors
pytesseract
Pillow
scikit-learn
pandas
gunicorn
